usrloc: new parameter to specify reply codes for keepalive handling
- comma separated list of reply codes or reply class to be considered
  valid for keepalive handling

+++ src/modules/usrloc/usrloc_mod.c
@@ -220,6 +220,8 @@ int ul_hash_size = 10;
 int ul_db_insert_null = 0;
 int ul_db_timer_clean = 0;
 
+char *ul_ka_reply_codes_str = "0";
+
 /* flags */
 unsigned int ul_nat_bflag = (unsigned int)-1;
 unsigned int ul_init_flag = 0;
@@ -301,6 +303,7 @@ static param_export_t params[] = {
 	{"ka_timeout", PARAM_INT, &ul_keepalive_timeout},
 	{"ka_loglevel", PARAM_INT, &ul_ka_loglevel},
 	{"ka_logmsg", PARAM_STR, &ul_ka_logmsg},
+	{"ka_reply_codes", PARAM_STRING, &ul_ka_reply_codes_str},
 	{"load_rank", PARAM_INT, &ul_load_rank},
 	{"db_clean_tcp", PARAM_INT, &ul_db_clean_tcp},
 	{0, 0, 0}
@@ -335,6 +338,9 @@ static int mod_init(void)
 	int i;
 	udomain_t *d;
 
+	if(ul_ka_parse_reply_codes(ul_ka_reply_codes_str)) {
+		return -1;
+	}
 	if(ul_rm_expired_delay != 0) {
 		if(ul_db_mode != DB_ONLY) {
 			LM_ERR("rm expired delay feature is available for db only mode\n");
